Added Resource Access and Custom Authorization Policies

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,5 +1,8 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
@@ -12,15 +15,17 @@ public class RestaurantService : IRestaurantService
     private readonly RestaurantDbContext _dbContext;
     private readonly IMapper _mapper;
     private readonly ILogger<RestaurantService> _logger;
+    private readonly IAuthorizationService _authorizationService;
 
-    public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+    public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService)
     {
         _dbContext = dbContext;
         _mapper = mapper;
         _logger = logger;
+        _authorizationService = authorizationService;
     }
 
-    public void Delete(int id)
+    public void Delete(int id, ClaimsPrincipal user)
     {
         _logger.LogError($"Restaurant with id: {id} DELETE action invoked");
         
@@ -30,13 +35,18 @@ public class RestaurantService : IRestaurantService
         
         if (restaurant is null) throw new NotFoundException("Restaurant not found");
         
+        var authoizationResult = _authorizationService
+            .AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+        if(!authoizationResult.Succeeded) throw new ForbidException();
+        
         _dbContext.Restaurants.Remove(restaurant);
         _dbContext.SaveChanges();
     }
     
-    public int Create(CreateRestaurantDto dto)
+    public int Create(CreateRestaurantDto dto, int userId)
     {
         var restaurant = _mapper.Map<Restaurant>(dto);
+        restaurant.CreatedById = userId;
         _dbContext.Restaurants.Add(restaurant);
         _dbContext.SaveChanges();
 
@@ -69,7 +79,7 @@ public class RestaurantService : IRestaurantService
         return result;
     }
     
-    public void Update(int id, UpdateRestaurantDto dto)
+    public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
     {
         var restaurant = _dbContext
             .Restaurants
@@ -77,6 +87,10 @@ public class RestaurantService : IRestaurantService
 
         if (restaurant is null) throw new NotFoundException("Restaurant not found");
 
+        var authoizationResult = _authorizationService
+            .AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+        if(!authoizationResult.Succeeded) throw new ForbidException();
+        
         restaurant.Name = dto.Name;
         restaurant.Description = dto.Description;
         restaurant.HasDelivery = dto.HasDelivery;
